// src/app/swagger-models/model/book.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Book { 
    authors?: string;
    description?: string;
    id?: number;
    imageUrl?: string;
    isbn?: string;
    numberOfPages?: number;
    price?: number;
    publicationYear?: number;
    publisher?: string;
    subtitle?: string;
    title?: string;
}